requirements.txt
bcrypt==4.2.0
black==24.8.0
boto==2.49.0
boto3==1.35.34
botocore==1.35.34
cffi==1.17.1
click==8.1.7
coverage==7.6.1
cryptography==43.0.1
exceptiongroup==1.2.2
iniconfig==2.0.0
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==24.1
paramiko==3.5.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3>=1.26.8
paho-mqtt==2.1.0